More Tree
Quite a lot of tree.
Need to design a string parser for the tree creation. The string should include the information regarding the type of operations that the data will be subverted to, the filtering criteria and the data to be stored.
Also split tree_types in a different json.